At least one conversion works
High point so far:

assert.ok(close(enunit(1, 'gallon/hour').as('cm^3/second'), 1.05150327));

I introduced the idea of a 'UnitSpace', which is like a namespace in that
it prevents units you register from polluting the global converter. Base
and derived units can be registered in a UnitSpace. There is a built-in
one with some of the normal things like grams, miles, and gallons.
Arithmetic operators are mostly implemented but not tested yet.