UnitValue prototype cleaner; no more unitSpace
UnitValue did have its unitSpace attached to its prototype so that it
could check that instances from different UnitSpaces did not interact.
That is not actually necessary, though; it does just as well to check that
the two functions of the same name (I used 'as') on the two UnitValues
match.

Now only things that the module user ought to use are part of UnitValues'
prototypes.